+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 ### Added
 - Add commands to query the word without modification.
+- Add the configuration to enable or disable 'Show Call Hierarchy'.
 
 ### Fixed
 - Fix the exception when the cscope returns the full path for result.